+++ src/exchange.ts
@@ -8,6 +8,7 @@ import {
   AccountInfo,
   AccountMeta,
   Commitment,
+  Context,
 } from "@solana/web3.js";
 import * as utils from "./utils";
 import * as constants from "./constants";
@@ -244,6 +245,16 @@ export class Exchange {
   private _clockSubscriptionId: number;
 
   /**
+   * The subscription id for the pricing account.
+   */
+  private _pricingSubscriptionId: number = undefined;
+
+  /**
+   * The subscription id for the state account.
+   */
+  private _stateSubscriptionId: number = undefined;
+
+  /**
    * @param interval   How often to poll zeta group and state in seconds.
    */
   public get pollInterval(): number {
@@ -297,6 +308,9 @@ export class Exchange {
   }
   private _useAutoPriorityFee: boolean = false;
 
+  private _autoPriorityFeeOffset: number = 0;
+  private _autoPriorityFeeMultiplier: number = 1;
+
   // Micro lamports per CU of fees.
   public get autoPriorityFeeUpperLimit(): number {
     return this._autoPriorityFeeUpperLimit;
@@ -311,6 +325,11 @@ export class Exchange {
 
   public toggleAutoPriorityFee() {
     this._useAutoPriorityFee = !this._useAutoPriorityFee;
+  }
+
+  public setAutoPriorityFeeScaling(offset: number = 0, multiplier: number = 1) {
+    this._autoPriorityFeeMultiplier = multiplier;
+    this._autoPriorityFeeOffset = offset;
   }
 
   public updatePriorityFee(microLamportsPerCU: number) {
@@ -542,7 +561,7 @@ export class Exchange {
   public async load(
     loadConfig: types.LoadExchangeConfig,
     wallet = new types.DummyWallet(),
-    callback?: (asset: Asset, event: EventType, data: any) => void
+    callback?: (asset: Asset, event: EventType, slot: number, data: any) => void
   ) {
     if (this.isInitialized) {
       throw "Exchange already loaded";
@@ -690,7 +709,12 @@ export class Exchange {
         .map((obj) => obj.prioritizationFee); // Take a list of prioritizationFee values only
 
       let median = utils.median(fees);
-      this._priorityFee = Math.min(median, this._autoPriorityFeeUpperLimit);
+      let medianScaled =
+        this._autoPriorityFeeOffset + median * this._autoPriorityFeeMultiplier;
+      this._priorityFee = Math.min(
+        medianScaled,
+        this._autoPriorityFeeUpperLimit
+      );
       console.log(
         `AutoUpdate priority fee. New fee = ${this._priorityFee} microlamports per compute unit`
       );
@@ -701,16 +725,16 @@ export class Exchange {
 
   private async subscribeOracle(
     assets: Asset[],
-    callback?: (asset: Asset, type: EventType, data: any) => void
+    callback?: (asset: Asset, type: EventType, slot: number, data: any) => void
   ) {
     return this._oracle.subscribePriceFeeds(
       assets,
-      (asset: Asset, price: OraclePrice) => {
+      (asset: Asset, price: OraclePrice, slot: number) => {
         if (this.isInitialized) {
           this._riskCalculator.updateMarginRequirements(asset);
         }
         if (callback !== undefined) {
-          callback(asset, EventType.ORACLE, price);
+          callback(asset, EventType.ORACLE, slot, price);
         }
       }
     );
@@ -723,14 +747,14 @@ export class Exchange {
 
   private subscribeClock(
     clockData: types.ClockData,
-    callback?: (asset: Asset, type: EventType, data: any) => void
+    callback?: (asset: Asset, type: EventType, slot: number, data: any) => void
   ) {
     if (this._clockSubscriptionId !== undefined) {
       throw Error("Clock already subscribed to.");
     }
     this._clockSubscriptionId = this.provider.connection.onAccountChange(
       SYSVAR_CLOCK_PUBKEY,
-      async (accountInfo: AccountInfo<Buffer>, _context: any) => {
+      async (accountInfo: AccountInfo<Buffer>, context: any) => {
         this.setClockData(utils.getClockData(accountInfo));
 
         await Promise.all(
@@ -740,7 +764,7 @@ export class Exchange {
         );
 
         if (callback !== undefined) {
-          callback(null, EventType.CLOCK, null);
+          callback(null, EventType.CLOCK, context.slot, null);
         }
         try {
           if (
@@ -803,39 +827,39 @@ export class Exchange {
   }
 
   private subscribeState(
-    callback?: (asset: Asset, type: EventType, data: any) => void
+    callback?: (asset: Asset, type: EventType, slot: number, data: any) => void
   ) {
-    let eventEmitter = this._program.account.state.subscribe(
+    this._stateSubscriptionId = this.connection.onAccountChange(
       this._stateAddress,
-      this.provider.connection.commitment
-    );
+      async (accountInfo: AccountInfo<Buffer>, context: Context) => {
+        this._state = this.program.coder.accounts.decode(
+          "State",
+          accountInfo.data
+        );
 
-    eventEmitter.on("change", async (state: State) => {
-      this._state = state;
-      if (callback !== undefined) {
-        callback(null, EventType.EXCHANGE, null);
+        if (callback !== undefined) {
+          callback(null, EventType.EXCHANGE, context.slot, null);
+        }
       }
-    });
-
-    this._eventEmitters.push(eventEmitter);
+    );
   }
 
   private subscribePricing(
-    callback?: (asset: Asset, type: EventType, data: any) => void
+    callback?: (asset: Asset, type: EventType, slot: number, data: any) => void
   ) {
-    let eventEmitter = this._program.account.pricing.subscribe(
+    this._pricingSubscriptionId = this.connection.onAccountChange(
       this._pricingAddress,
-      this.provider.connection.commitment
-    );
+      async (accountInfo: AccountInfo<Buffer>, context: Context) => {
+        this._pricing = this.program.coder.accounts.decode(
+          "Pricing",
+          accountInfo.data
+        );
 
-    eventEmitter.on("change", async (pricing: Pricing) => {
-      this._pricing = pricing;
-      if (callback !== undefined) {
-        callback(null, EventType.PRICING, null);
+        if (callback !== undefined) {
+          callback(null, EventType.PRICING, context.slot, null);
+        }
       }
-    });
-
-    this._eventEmitters.push(eventEmitter);
+    );
   }
 
   public getZetaGroupMarkets(asset: Asset): ZetaGroupMarkets {
@@ -1085,8 +1109,19 @@ export class Exchange {
       this._clockSubscriptionId = undefined;
     }
 
-    await this._program.account.pricing.unsubscribe(this._pricingAddress);
-    await this._program.account.state.unsubscribe(this._stateAddress);
+    if (this._pricingSubscriptionId !== undefined) {
+      await this._provider.connection.removeAccountChangeListener(
+        this._pricingSubscriptionId
+      );
+      this._pricingSubscriptionId = undefined;
+    }
+
+    if (this._stateSubscriptionId !== undefined) {
+      await this._provider.connection.removeAccountChangeListener(
+        this._stateSubscriptionId
+      );
+      this._stateSubscriptionId = undefined;
+    }
 
     for (var i = 0; i < this._programSubscriptionIds.length; i++) {
       await this.connection.removeProgramAccountChangeListener(